feat(front): move token types to external file
