+++ src/front/ast_creator/lexer.rs
@@ -1,55 +1,7 @@
 use std::collections::VecDeque;
 use std::io::Read;
-use std::str::{CharIndices};
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum TokenKind {
-    // import
-    Use,
-    DoubleColon,
-
-    Ident(String),
-
-    // literals
-    LNull,
-    LInt(i32),
-
-    // type keyword
-    TVoid,
-    TInt,
-
-    // definition declaration
-    Static,
-    Struct,
-    Fn,
-
-    // misc
-    Colon,
-    SemiColon,
-    Comma,
-
-    LBrace,
-    RBrace,
-
-    Eof,
-}
-
-#[derive(Debug)]
-pub enum TokenError {
-    InvalidToken(String),
-    MultipleDecimals,
-    Unknown,
-}
-
-pub struct Span {
-    pub lo: usize,
-    pub hi: usize,
-}
-
-pub struct Token {
-    pub kind: TokenKind,
-    pub span: Span,
-}
+use std::str::CharIndices;
+use crate::front::ast_creator::token_types::{Span, Token, TokenError, TokenKind};
 
 pub struct Lexer<'src> {
     src: &'src str,
